requirements.txt
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.3.0
defusedxml==0.6.0
dj-stripe==2.3.0
Django==2.2.8
django-allauth==0.41.0
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.5.0
django-crispy-forms==1.9.0
django-debug-toolbar==1.10.1
django-filter==2.3.0
django-jquery==3.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
future==0.16.0
idna==2.8
ipython==5.8.0
ipython-genutils==0.2.0
jsonfield==3.1.0
mysql-connector-python==8.0.18
oauthlib==3.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pkg-resources==0.0.0
progressbar2==3.50.1
prompt-toolkit==1.0.18
protobuf==3.12.2
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
python-dateutil==2.7.3
python-utils==2.4.0
python3-openid==3.1.0
pytz==2020.1
random-password-generator==2.1.0
requests==2.21.0
requests-oauthlib==1.3.0
simplegeneric==0.8.1
sip==4.19.8
six==1.12.0
sqlparse==0.2.4
stripe==2.45.0
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.24.3
virtualenv==16.5.0
wcwidth==0.2.4
